float struggle is not that bad anymore

+++ libft/ft_pow.c
@@ -6,7 +6,7 @@
 /*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/10 10:53:47 by mviinika          #+#    #+#             */
-/*   Updated: 2022/03/24 13:53:12 by mviinika         ###   ########.fr       */
+/*   Updated: 2022/03/24 19:45:37 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 long double	ft_pow(long double num, long double pow)
 {
-	double	res;
+	long double	res;
 	int		counter;
 
 	res = 1;
